+++ cdp-agentkit-core/typescript/src/actions/cdp/index.ts
@@ -13,7 +13,7 @@ import { TransferNftAction } from "./transfer_nft";
 import { WrapEthAction } from "./wrap_eth";
 
 // TODO: ask John if this is the right path for morpho actions
-import { MORPHO_ACTIONS } from "./morpho";
+import { MORPHO_ACTIONS } from "./defi/morpho";
 import { PYTH_ACTIONS } from "./data/pyth";
 import { WOW_ACTIONS } from "./defi/wow";
