+++ cdp-agentkit-core/typescript/src/tests/defi_morpho_deposit_test.ts
@@ -1,19 +1,26 @@
-import { Coinbase, ContractInvocation, Wallet } from "@coinbase/coinbase-sdk";
-import { MorphoWithdrawAction } from "../actions/cdp/morpho/withdraw";
-import { METAMORPHO_ABI } from "../actions/cdp/morpho/constants";
+import { Coinbase, ContractInvocation, Wallet, Asset } from "@coinbase/coinbase-sdk";
+import { MorphoDepositAction } from "../actions/cdp/defi/morpho/deposit";
+import { METAMORPHO_ABI } from "../actions/cdp/defi/morpho/constants";
+import { approve } from "../actions/cdp/defi/morpho/utils";
+import { Decimal } from "decimal.js";
 
 const MOCK_VAULT_ADDRESS = "0x1234567890123456789012345678901234567890";
-const MOCK_ASSETS = "1000000000000000000"; // 1 token in wei
+const MOCK_ASSETS = "1000000000000000000";
 const MOCK_RECEIVER_ID = "0x9876543210987654321098765432109876543210";
+const MOCK_TOKEN_ADDRESS = "0x4200000000000000000000000000000000000006";
 
-describe("Morpho Withdraw Input", () => {
-  const action = new MorphoWithdrawAction();
+jest.mock("../actions/cdp/defi/morpho/utils");
+const mockApprove = approve as jest.MockedFunction<typeof approve>;
+
+describe("Morpho Deposit Input", () => {
+  const action = new MorphoDepositAction();
 
   it("should successfully parse valid input", () => {
     const validInput = {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
     const result = action.argsSchema.safeParse(validInput);
@@ -29,18 +36,16 @@ describe("Morpho Withdraw Input", () => {
     expect(result.success).toBe(false);
   });
 
-  it("should fail with invalid Vault address", () => {
+  it("should fail with invalid vault address", () => {
     const invalidInput = {
       vaultAddress: "not_an_address",
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
     const result = action.argsSchema.safeParse(invalidInput);
 
     expect(result.success).toBe(false);
-    if (!result.success) {
-      expect(result.error.issues[0].path[0]).toBe("vaultAddress");
-    }
   });
 
   it("should handle valid asset string formats", () => {
@@ -48,9 +53,14 @@ describe("Morpho Withdraw Input", () => {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
-    const validInputs = [{ ...validInput, assets: "1000000000000000000" }];
+    const validInputs = [
+      { ...validInput, assets: "1000000000000000000" },
+      { ...validInput, assets: "1.5" },
+      { ...validInput, assets: "0.00001" },
+    ];
 
     validInputs.forEach(input => {
       const result = action.argsSchema.safeParse(input);
@@ -63,6 +73,7 @@ describe("Morpho Withdraw Input", () => {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
     const invalidInputs = [
@@ -79,12 +90,12 @@ describe("Morpho Withdraw Input", () => {
   });
 });
 
-describe("Morpho Withdraw Action", () => {
+describe("Morpho Deposit Action", () => {
   const NETWORK_ID = Coinbase.networks.BaseSepolia;
   const TRANSACTION_HASH = "0xabcdef1234567890";
   const TRANSACTION_LINK = `https://etherscan.io/tx/${TRANSACTION_HASH}`;
 
-  const action = new MorphoWithdrawAction();
+  const action = new MorphoDepositAction();
 
   let mockContractInvocation: jest.Mocked<ContractInvocation>;
   let mockWallet: jest.Mocked<Wallet>;
@@ -92,10 +103,8 @@ describe("Morpho Withdraw Action", () => {
   beforeEach(() => {
     mockContractInvocation = {
       wait: jest.fn().mockResolvedValue({
-        getTransaction: jest.fn().mockReturnValue({
-          getTransactionHash: jest.fn().mockReturnValue(TRANSACTION_HASH),
-          getTransactionLink: jest.fn().mockReturnValue(TRANSACTION_LINK),
-        }),
+        getTransactionHash: jest.fn().mockReturnValue(TRANSACTION_HASH),
+        getTransactionLink: jest.fn().mockReturnValue(TRANSACTION_LINK),
       }),
     } as unknown as jest.Mocked<ContractInvocation>;
 
@@ -108,47 +117,82 @@ describe("Morpho Withdraw Action", () => {
     } as unknown as jest.Mocked<Wallet>;
 
     mockWallet.invokeContract.mockResolvedValue(mockContractInvocation);
+
+    jest.spyOn(Asset, "fetch").mockResolvedValue({
+      toAtomicAmount: jest.fn().mockImplementation((amount: Decimal) => amount.toString()),
+    } as unknown as Asset);
+
+    mockApprove.mockResolvedValue("Approval successful");
   });
 
-  it("should successfully withdraw from Morpho Vault", async () => {
+  it("should successfully deposit to Morpho vault", async () => {
     const args = {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
+
+    const assets = BigInt(MOCK_ASSETS);
 
     const response = await action.func(mockWallet, args);
 
+    expect(mockApprove).toHaveBeenCalledWith(
+      mockWallet,
+      MOCK_TOKEN_ADDRESS,
+      MOCK_VAULT_ADDRESS,
+      assets,
+    );
+
     expect(mockWallet.invokeContract).toHaveBeenCalledWith({
       contractAddress: MOCK_VAULT_ADDRESS,
-      method: "withdraw",
+      method: "deposit",
       abi: METAMORPHO_ABI,
       args: {
         assets: MOCK_ASSETS,
         receiver: MOCK_RECEIVER_ID,
-        owner: MOCK_RECEIVER_ID,
       },
     });
+
     expect(mockContractInvocation.wait).toHaveBeenCalled();
-    expect(response).toContain(`Withdrawn ${MOCK_ASSETS}`);
-    expect(response).toContain(`from Morpho Vault ${MOCK_VAULT_ADDRESS}`);
+    expect(response).toContain(`Deposited ${MOCK_ASSETS}`);
+    expect(response).toContain(`to Morpho Vault ${MOCK_VAULT_ADDRESS}`);
     expect(response).toContain(`with transaction hash: ${TRANSACTION_HASH}`);
     expect(response).toContain(`and transaction link: ${TRANSACTION_LINK}`);
   });
 
-  it("should handle errors when withdrawing", async () => {
+  it("should handle approval failure", async () => {
     const args = {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
-    const error = new Error("API Error");
+    mockApprove.mockResolvedValue("Error: Approval failed");
+
+    const response = await action.func(mockWallet, args);
+
+    expect(mockApprove).toHaveBeenCalled();
+    expect(response).toContain("Error approving Morpho Vault as spender: Error: Approval failed");
+    expect(mockWallet.invokeContract).not.toHaveBeenCalled();
+  });
+
+  it("should handle deposit errors", async () => {
+    const args = {
+      vaultAddress: MOCK_VAULT_ADDRESS,
+      assets: MOCK_ASSETS,
+      receiver: MOCK_RECEIVER_ID,
+      tokenAddress: MOCK_TOKEN_ADDRESS,
+    };
+
+    const error = new Error("Failed to deposit to Morpho vault");
     mockWallet.invokeContract.mockRejectedValue(error);
 
     const response = await action.func(mockWallet, args);
 
+    expect(mockApprove).toHaveBeenCalled();
     expect(mockWallet.invokeContract).toHaveBeenCalled();
-    expect(response).toContain(`Error withdrawing from Morpho Vault: ${error}`);
+    expect(response).toContain(`Error depositing to Morpho Vault: ${error}`);
   });
 });
