+++ cdp-agentkit-core/typescript/src/tests/defi_morpho_withdraw_test.ts
@@ -1,26 +1,19 @@
-import { Coinbase, ContractInvocation, Wallet, Asset } from "@coinbase/coinbase-sdk";
-import { MorphoDepositAction } from "../actions/cdp/morpho/deposit";
-import { METAMORPHO_ABI } from "../actions/cdp/morpho/constants";
-import { approve } from "../actions/cdp/morpho/utils";
-import { Decimal } from "decimal.js";
+import { Coinbase, ContractInvocation, Wallet } from "@coinbase/coinbase-sdk";
+import { MorphoWithdrawAction } from "../actions/cdp/defi/morpho/withdraw";
+import { METAMORPHO_ABI } from "../actions/cdp/defi/morpho/constants";
 
 const MOCK_VAULT_ADDRESS = "0x1234567890123456789012345678901234567890";
-const MOCK_ASSETS = "1000000000000000000";
+const MOCK_ASSETS = "1000000000000000000"; // 1 token in wei
 const MOCK_RECEIVER_ID = "0x9876543210987654321098765432109876543210";
-const MOCK_TOKEN_ADDRESS = "0x4200000000000000000000000000000000000006";
 
-jest.mock("../actions/cdp/morpho/utils");
-const mockApprove = approve as jest.MockedFunction<typeof approve>;
-
-describe("Morpho Deposit Input", () => {
-  const action = new MorphoDepositAction();
+describe("Morpho Withdraw Input", () => {
+  const action = new MorphoWithdrawAction();
 
   it("should successfully parse valid input", () => {
     const validInput = {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
     const result = action.argsSchema.safeParse(validInput);
@@ -36,16 +29,18 @@ describe("Morpho Deposit Input", () => {
     expect(result.success).toBe(false);
   });
 
-  it("should fail with invalid vault address", () => {
+  it("should fail with invalid Vault address", () => {
     const invalidInput = {
       vaultAddress: "not_an_address",
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
     const result = action.argsSchema.safeParse(invalidInput);
 
     expect(result.success).toBe(false);
+    if (!result.success) {
+      expect(result.error.issues[0].path[0]).toBe("vaultAddress");
+    }
   });
 
   it("should handle valid asset string formats", () => {
@@ -53,14 +48,9 @@ describe("Morpho Deposit Input", () => {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
-    const validInputs = [
-      { ...validInput, assets: "1000000000000000000" },
-      { ...validInput, assets: "1.5" },
-      { ...validInput, assets: "0.00001" },
-    ];
+    const validInputs = [{ ...validInput, assets: "1000000000000000000" }];
 
     validInputs.forEach(input => {
       const result = action.argsSchema.safeParse(input);
@@ -73,7 +63,6 @@ describe("Morpho Deposit Input", () => {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
     const invalidInputs = [
@@ -90,12 +79,12 @@ describe("Morpho Deposit Input", () => {
   });
 });
 
-describe("Morpho Deposit Action", () => {
+describe("Morpho Withdraw Action", () => {
   const NETWORK_ID = Coinbase.networks.BaseSepolia;
   const TRANSACTION_HASH = "0xabcdef1234567890";
   const TRANSACTION_LINK = `https://etherscan.io/tx/${TRANSACTION_HASH}`;
 
-  const action = new MorphoDepositAction();
+  const action = new MorphoWithdrawAction();
 
   let mockContractInvocation: jest.Mocked<ContractInvocation>;
   let mockWallet: jest.Mocked<Wallet>;
@@ -103,8 +92,10 @@ describe("Morpho Deposit Action", () => {
   beforeEach(() => {
     mockContractInvocation = {
       wait: jest.fn().mockResolvedValue({
-        getTransactionHash: jest.fn().mockReturnValue(TRANSACTION_HASH),
-        getTransactionLink: jest.fn().mockReturnValue(TRANSACTION_LINK),
+        getTransaction: jest.fn().mockReturnValue({
+          getTransactionHash: jest.fn().mockReturnValue(TRANSACTION_HASH),
+          getTransactionLink: jest.fn().mockReturnValue(TRANSACTION_LINK),
+        }),
       }),
     } as unknown as jest.Mocked<ContractInvocation>;
 
@@ -117,82 +108,47 @@ describe("Morpho Deposit Action", () => {
     } as unknown as jest.Mocked<Wallet>;
 
     mockWallet.invokeContract.mockResolvedValue(mockContractInvocation);
-
-    jest.spyOn(Asset, "fetch").mockResolvedValue({
-      toAtomicAmount: jest.fn().mockImplementation((amount: Decimal) => amount.toString()),
-    } as unknown as Asset);
-
-    mockApprove.mockResolvedValue("Approval successful");
   });
 
-  it("should successfully deposit to Morpho vault", async () => {
+  it("should successfully withdraw from Morpho Vault", async () => {
     const args = {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
-
-    const assets = BigInt(MOCK_ASSETS);
 
     const response = await action.func(mockWallet, args);
 
-    expect(mockApprove).toHaveBeenCalledWith(
-      mockWallet,
-      MOCK_TOKEN_ADDRESS,
-      MOCK_VAULT_ADDRESS,
-      assets,
-    );
-
     expect(mockWallet.invokeContract).toHaveBeenCalledWith({
       contractAddress: MOCK_VAULT_ADDRESS,
-      method: "deposit",
+      method: "withdraw",
       abi: METAMORPHO_ABI,
       args: {
         assets: MOCK_ASSETS,
         receiver: MOCK_RECEIVER_ID,
+        owner: MOCK_RECEIVER_ID,
       },
     });
-
     expect(mockContractInvocation.wait).toHaveBeenCalled();
-    expect(response).toContain(`Deposited ${MOCK_ASSETS}`);
-    expect(response).toContain(`to Morpho Vault ${MOCK_VAULT_ADDRESS}`);
+    expect(response).toContain(`Withdrawn ${MOCK_ASSETS}`);
+    expect(response).toContain(`from Morpho Vault ${MOCK_VAULT_ADDRESS}`);
     expect(response).toContain(`with transaction hash: ${TRANSACTION_HASH}`);
     expect(response).toContain(`and transaction link: ${TRANSACTION_LINK}`);
   });
 
-  it("should handle approval failure", async () => {
+  it("should handle errors when withdrawing", async () => {
     const args = {
       vaultAddress: MOCK_VAULT_ADDRESS,
       assets: MOCK_ASSETS,
       receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
     };
 
-    mockApprove.mockResolvedValue("Error: Approval failed");
-
-    const response = await action.func(mockWallet, args);
-
-    expect(mockApprove).toHaveBeenCalled();
-    expect(response).toContain("Error approving Morpho Vault as spender: Error: Approval failed");
-    expect(mockWallet.invokeContract).not.toHaveBeenCalled();
-  });
-
-  it("should handle deposit errors", async () => {
-    const args = {
-      vaultAddress: MOCK_VAULT_ADDRESS,
-      assets: MOCK_ASSETS,
-      receiver: MOCK_RECEIVER_ID,
-      tokenAddress: MOCK_TOKEN_ADDRESS,
-    };
-
-    const error = new Error("Failed to deposit to Morpho vault");
+    const error = new Error("API Error");
     mockWallet.invokeContract.mockRejectedValue(error);
 
     const response = await action.func(mockWallet, args);
 
-    expect(mockApprove).toHaveBeenCalled();
     expect(mockWallet.invokeContract).toHaveBeenCalled();
-    expect(response).toContain(`Error depositing to Morpho Vault: ${error}`);
+    expect(response).toContain(`Error withdrawing from Morpho Vault: ${error}`);
   });
 });
